feat(brig): `brig run` as a go library
You can write any code that reuses `brig run`, to delegate running distributed scripts onto brigade, but with a more efficient and type-safe way.

This is a backward-compatible change, irrelevant to the feature of `brig run` and `brig rerun` commands.

This tweaks the original implementation of `brig run`, removing global variable references, extracting almost all the logics of it into the brand-new `brigade/script` go pkg.
So that it can be imported by other golang projects.

The script runner is now named `DelegatedRunner`, to make it extra clear to users, that the runner delegates a script to brigade, rather than running it in place.

The basic usage of this pkg, and how the body of `brig run` look like now, is:

```
runner, err := script.NewDelegatedRunner(kc, destination, globalNamespace, runNoProgress, runBackground, globalVerbose)
if err != nil {
    return err
}

runner.SendScript(proj, script, event, commitish, ref, payload, logLevel)

runner.SendBuild(build)
```

Resolves #622
